Improve WordImporter #41
If there is a UC without any normal sequence description (but with valid
header - name and id) a graph with two nodes (start node and end node)
was visualized. This could be the case if a user just want to prepare a
use case (description,...) but not specify normal sequences yet.
From now on the user will be shown a warning that normal sequence was
not specified, but all general UC attributes (Description,
Precondition,...) will still be shown in our tool.